+++ packages/article-label/__tests__/web/jest.config.js
@@ -1,3 +1,5 @@
 const jestConfigurator = require("@times-components/jest-configurator");
 
-module.exports = jestConfigurator("article-label", "web");
+module.exports = Object.assign(jestConfigurator("article-label", "web"), {
+  collectCoverageFrom: ["**/packages/article-label/article-label.js"]
+});