+++ packages/article/__tests__/web/article.web.test.js
@@ -1,6 +1,24 @@
 /* eslint-env jest */
+import "react-native";
+import React from "react";
+import renderer from "react-test-renderer";
+import Article from "../../article";
 import shared from "../shared";
+
+const articleFixtureNoLeadAsset = require("../../fixtures/no-lead-asset.json");
 
 describe("Article test on web", () => {
   shared();
+
+  it("renders article with no lead asset", () => {
+    const tree = renderer
+      .create(
+        <Article
+          {...articleFixtureNoLeadAsset.data}
+          analyticsStream={() => {}}
+        />
+      )
+      .toJSON();
+    expect(tree).toMatchSnapshot();
+  });
 });