+++ packages/card/__tests__/android/jest.config.js
@@ -1,3 +1,10 @@
 const jestConfigurator = require("@times-components/jest-configurator");
 
-module.exports = jestConfigurator("card", "android");
+module.exports = Object.assign(jestConfigurator("card", "android"), {
+  collectCoverageFrom: [
+    "**/packages/card/*.js",
+    "!**/packages/card/*stories*",
+    "!**/packages/card/*.web.js",
+    "!**/packages/card/__test__/**"
+  ]
+});