+++ packages/error-view/__tests__/throws-error.js
@@ -1,0 +1,11 @@
+/* eslint-disable react/prefer-stateless-function, react/require-render-return */
+
+import { Component } from "react";
+
+class Erroring extends Component {
+  render() {
+    throw new Error("an error occurred");
+  }
+}
+
+export default Erroring;