+++ packages/article-flag/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.9.2"></a>
+## [0.9.2](https://github.com/newsuk/times-components/compare/@times-components/article-flag@0.9.1...@times-components/article-flag@0.9.2) (2017-12-14)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-flag
+
+<a name="0.9.1"></a>
+## [0.9.1](https://github.com/newsuk/times-components/compare/@times-components/article-flag@0.9.0...@times-components/article-flag@0.9.1) (2017-12-13)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-flag
+
 <a name="0.9.0"></a>
 # [0.9.0](https://github.com/newsuk/times-components/compare/@times-components/article-flag@0.8.5...@times-components/article-flag@0.9.0) (2017-12-12)