+++ packages/brightcove-video/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.9.1"></a>
+## [1.9.1](https://github.com/newsuk/times-components/compare/@times-components/brightcove-video@1.9.0...@times-components/brightcove-video@1.9.1) (2017-12-14)
+
+
+
+
+**Note:** Version bump only for package @times-components/brightcove-video
+
 <a name="1.9.0"></a>
 # [1.9.0](https://github.com/newsuk/times-components/compare/@times-components/brightcove-video@1.8.4...@times-components/brightcove-video@1.9.0) (2017-12-12)