+++ packages/markup/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.14.2"></a>
+## [0.14.2](https://github.com/newsuk/times-components/compare/@times-components/markup@0.14.1...@times-components/markup@0.14.2) (2017-12-14)
+
+
+
+
+**Note:** Version bump only for package @times-components/markup
+
+<a name="0.14.1"></a>
+## [0.14.1](https://github.com/newsuk/times-components/compare/@times-components/markup@0.14.0...@times-components/markup@0.14.1) (2017-12-13)
+
+
+
+
+**Note:** Version bump only for package @times-components/markup
+
 <a name="0.14.0"></a>
 # [0.14.0](https://github.com/newsuk/times-components/compare/@times-components/markup@0.13.4...@times-components/markup@0.14.0) (2017-12-12)