+++ packages/utils/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.0.7"></a>
+## [0.0.7](https://github.com/newsuk/times-components/compare/@times-components/utils@0.0.6...@times-components/utils@0.0.7) (2017-12-14)
+
+
+
+
+**Note:** Version bump only for package @times-components/utils
+
 <a name="0.0.6"></a>
 ## [0.0.6](https://github.com/newsuk/times-components/compare/@times-components/utils@0.0.5...@times-components/utils@0.0.6) (2017-12-12)